backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Pydantic & Core Tools
pydantic>=2.5.0
python-dotenv==1.0.0

# HTTP Clients
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.9.1

# Transcript Extraction
youtube-transcript-api==0.6.2
yt-dlp

# Chapter Generation
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Text Processing
tiktoken>=0.5.2

# SSL fix
certifi

# NEW (for ASR fallback)
faster-whisper>=1.0.0

# Optional but recommended (if using ffmpeg via python)
ffmpeg-python>=0.2.0

# Database
motor>=3.3.2
pymongo>=4.6.0

# Auth (JWT + bcrypt + email validation)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.5.0

# Certificate PDF (optional — install manually if needed)
# reportlab>=4.0.0